+++ admin/changedata.php
@@ -108,6 +108,7 @@ if (isset($_POST['submitted'])) {
 		if(isset($_POST['post-content'])) 		{	$content = safe_slash_html($_POST['post-content']);	}
 		if (isset($_POST['post-component']))	{ $component = safe_slash_html($_POST['post-component']); }
 		if (isset($_POST['post-component-enable']))	{ $componentEnable = safe_slash_html($_POST['post-component-enable']); }
+		if (isset($_POST['post-lang'])) 			{ $lang = var_out(xss_clean($_POST['post-lang'])); }
 		if(isset($_POST['post-menu-order'])) 	{ 
 			if (is_numeric($_POST['post-menu-order'])) 
 			{
@@ -162,6 +163,7 @@ if (isset($_POST['submitted'])) {
 		$xml->addChild('private', $private);
 		$xml->addChild('author', $author);
 		$xml->addChild('lastAuthor', $USR);
+		$xml->addChild('lang', $lang);
 
 		exec_action('changedata-save');
 		if (isset($_POST['autosave']) && $_POST['autosave'] == 'true' && $autoSaveDraft == true) {