+++ admin/lang/en_US.php
@@ -602,6 +602,10 @@ $i18n = array(
 "ENABLE_CODE_ED"		=>	"Enable the code editor - <b style=\"font-weight:100\">Highlights code syntax and improves usability of editors for components and theme files</b>", // settings.php
 "MAINTENANCE_WARNING" => "Website is working in <strong>Maintenance Mode</strong> and not available for visitors",
 "ALLOW_ACCESS_IN_MAINTENANCE" => "Access to website in Maintenance Mode - <b style=\"font-weight:100\">Allow access to website while Maintenance Mode is enable</b>",
+"LABEL_WEBSITELANG" => "Website Language",
+"DISPLAY_WEBSITELANG" => "A code of language of website content (en, de, fr, it, ru, be, ua, etc.). Set per page language in page editor",
+"LABEL_PAGELANG"  => "Language",
+"PLACEHOLDER_LANG" => "en",
 );
 
 ?>